clean or compose
