+++ projects/compose/koin-compose-viewmodel/src/commonMain/kotlin/org/koin/compose/viewmodel/dsl/ViewModelOf.kt
@@ -43,6 +43,7 @@ import org.koin.core.module.dsl.onOptions
  *
  * @see new
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel> Module.viewModelOf(
     crossinline constructor: () -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -51,6 +52,7 @@ inline fun <reified R : ViewModel> Module.viewModelOf(
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1> Module.viewModelOf(
     crossinline constructor: (T1) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -59,6 +61,7 @@ inline fun <reified R : ViewModel, reified T1> Module.viewModelOf(
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1, reified T2> Module.viewModelOf(
     crossinline constructor: (T1, T2) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -67,6 +70,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2> Module.viewModelOf(
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -75,6 +79,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3> Module.vi
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -83,6 +88,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -91,6 +97,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -99,6 +106,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -107,6 +115,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -115,6 +124,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -123,6 +133,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -131,6 +142,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -139,6 +151,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -147,6 +160,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -155,6 +169,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -163,6 +178,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -171,6 +187,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -179,6 +196,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -187,6 +205,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -195,6 +214,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18, reified T19> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -203,6 +223,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18, reified T19, reified T20> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -211,6 +232,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18, reified T19, reified T20, reified T21> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20, T21) -> R,
     noinline options: DefinitionOptions<R>? = null,
@@ -219,6 +241,7 @@ inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T
 /**
  * @see viewModelOf
  */
+@Deprecated("use org.koin.viewmodel.dsl", ReplaceWith(expression = "viewModel", imports = ["org.koin.compose.dsl"]))
 inline fun <reified R : ViewModel, reified T1, reified T2, reified T3, reified T4, reified T5, reified T6, reified T7, reified T8, reified T9, reified T10, reified T11, reified T12, reified T13, reified T14, reified T15, reified T16, reified T17, reified T18, reified T19, reified T20, reified T21, reified T22> Module.viewModelOf(
     crossinline constructor: (T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20, T21, T22) -> R,
     noinline options: DefinitionOptions<R>? = null,
