+++ projects/core/koin-core-viewmodel/src/commonMain/kotlin/org/koin/viewmodel/BundleExt.kt
@@ -1,20 +1,4 @@
-///*
-// * Copyright 2017-2023 the original author or authors.
-// *
-// * Licensed under the Apache License, Version 2.0 (the "License");
-// * you may not use this file except in compliance with the License.
-// * You may obtain a copy of the License at
-// *
-// *      http://www.apache.org/licenses/LICENSE-2.0
-// *
-// * Unless required by applicable law or agreed to in writing, software
-// * distributed under the License is distributed on an "AS IS" BASIS,
-// * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// * See the License for the specific language governing permissions and
-// * limitations under the License.
-// */
-
-package org.koin.viewmodel.factory
+package org.koin.viewmodel
 
 import androidx.core.bundle.Bundle
 import androidx.lifecycle.DEFAULT_ARGS_KEY
@@ -25,7 +9,6 @@ import androidx.lifecycle.viewmodel.CreationExtras
 import androidx.lifecycle.viewmodel.MutableCreationExtras
 import androidx.savedstate.SavedStateRegistryOwner
 import org.koin.core.annotation.KoinInternalApi
-
 
 /**
  * Scope extensions to help for ViewModel
